Refactor Nextcloud compose files
Until now we kept the MariaDB as a separate docker compose service. The idea was to use the MariaDB as a shared database server for multiple services. However, we never really did it that way.
Subsequent applications all have come with their own database service.
So it makes sense to migrate the MariaDB into the Nextcloud docker compose definition.